Only reinstall bower deps on v0.15.0

+++ test-js/integration.js
@@ -32,11 +32,17 @@ const testBowerJson = {
     "url": "git://github.com/not-real/not-real.git"
   },
   "dependencies": {
+    "purescript-console": "^0.1.0",
+  }
+};
+
+const testBowerJsonV0_15_0 = Object.assign({}, testBowerJson, {
+  "dependencies": {
     "purescript-prelude": "^5.0.1",
     "purescript-console": "^5.0.0",
     "purescript-effect": "^3.0.0"
   }
-};
+});
 
 function resolvePath(cmd) {
   return new Promise((resolve, reject) => {
@@ -657,19 +663,23 @@ describe("integration tests", function() {
 
       yield fs.writeFile(path.join(temp, "bower.json"), JSON.stringify(testBowerJson));
 
-      // For tests to pass on the `v0.15.0-alpha-01` purs version,
-      // we need to use a custom fork of a non-core repo
-      // (i.e. working-group-purescript-es/purescript-prelude#es-modules-libraries).
-      // Since those repos' `bower.json` files point to dependencies
-      // whose "versions" are branch names as well, the `bower.json` JSON parsing fails.
-      // So, we need to remove the `bower_components` folder that gets set up
-      // when `pulp init` is called, and re-install the bower deps
-      // using the `bower.json` file.
-      //
-      // Furthermore, we have to install all 3 deps used in a typical `pulp init`
-      // setup. Otherwise, `pulp version` fails due to other dependency-related reasons.
-      yield sh("rm -rf bower_components");
-      yield sh("bower install");
+      const psVer = yield getPursVersion(sh, assert);
+      if (semver.gte(psVer, semver.parse('0.15.0'))) {
+        // For tests to pass on the `v0.15.0-alpha-01` purs version,
+        // we need to use a custom fork of a non-core repo
+        // (i.e. working-group-purescript-es/purescript-prelude#es-modules-libraries).
+        // Since those repos' `bower.json` files point to dependencies
+        // whose "versions" are branch names as well, the `bower.json` JSON parsing fails.
+        // So, we need to remove the `bower_components` folder that gets set up
+        // when `pulp init` is called, and re-install the bower deps
+        // using the `bower.json` file.
+        //
+        // Furthermore, we have to install all 3 deps used in a typical `pulp init`
+        // setup. Otherwise, `pulp version` fails due to other dependency-related reasons.
+        yield fs.writeFile(path.join(temp, "bower.json"), JSON.stringify(testBowerJsonV0_15_0));
+        yield sh("rm -rf bower_components");
+        yield sh("bower install");
+      }
 
       yield sh("git commit -am \"updating bower.json\"");
       yield sh("git tag v2.0.0");
